requirements.txt
pandas
numpy
scipy
openai
pymupdf
streamlit
nltk
python-dotenv